+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [1.0.4] - 2025-01-03
+
+### Fixed
+
+- Better support for module hierarchy.
+
 ## [1.0.3] - 2024-12-31
 
 - Improved ui for inlay hints.
@@ -28,7 +34,8 @@
 
 - Initial working version
 
-[Unreleased]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.3...HEAD
+[Unreleased]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.4...HEAD
+[1.0.4]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/hpalma/intellij-domain-translator/compare/v1.0.0...v1.0.1